+++ Models/Db/PipeModel.cs
@@ -16,9 +16,9 @@ namespace Pipe.Models.Db
         public long SteelId { get; set; }
         [ForeignKey("SteelId")]
         public Steel Steel { get; set; }
-        public int Diameter { get; set; }
-        public int Thickness { get; set; }
-        public int Length { get; set; }
-        public int Weight { get; set; }
+        public decimal Diameter { get; set; }
+        public decimal Thickness { get; set; }
+        public decimal Length { get; set; }
+        public decimal Weight { get; set; }
     }
 }
